backend/requirements.txt
fastapi
uvicorn
python-multipart
python-socketio
pydantic
python-dotenv
firebase-admin
PyJWT
passlib
Pillow
google-cloud-firestore
google-cloud-vision

twilio
face_recognition
requests
groq
